+++ src/app/dashboard/meetings/meet/page.jsx
@@ -1,36 +1,36 @@
 "use client";
 
-import { useEffect } from "react";
-import { selectIsConnectedToRoom, selectRoomState, useHMSActions, useHMSStore } from "@100mslive/react-sdk";
+// import { useEffect } from "react";
+// import { selectIsConnectedToRoom, selectRoomState, useHMSActions, useHMSStore } from "@100mslive/react-sdk";
 
-import Conference from './components/Conference';
-import Footer from './components/Footer';
+// import Conference from './components/Conference';
+// import Footer from './components/Footer';
 import JoinForm from './components/JoinForm';
 
 const page = () => {
-    const isConnected = useHMSStore(selectIsConnectedToRoom);
-    const hmsActions = useHMSActions();
+    // const isConnected = useHMSStore(selectIsConnectedToRoom);
+    // const hmsActions = useHMSActions();
 
 
-    useEffect(() => {
-        window.onunload = () => {
-            if (isConnected) {
-                hmsActions.leave();
-            }
-        };
-    }, [hmsActions, isConnected]);
+    // useEffect(() => {
+    //     window.onunload = () => {
+    //         if (isConnected) {
+    //             hmsActions.leave();
+    //         }
+    //     };
+    // }, [hmsActions, isConnected]);
 
     return (
-        <div className="App">
-            {isConnected ? (
-                <>
-                    <Conference />
-                    <Footer />
-                </>
-            ) : (
+        // <div className=" bg-white">
+        //     {isConnected ? (
+        //         <>
+        //             {/* <Conference />
+        //             <Footer /> */}
+        //         </>
+        //     ) : (
                 <JoinForm />
-            )}
-        </div>
+        //     )}
+        // </div>
     );
 };
